+++ src/ColumnGroup.js
@@ -2,9 +2,8 @@ import { Component } from 'react';
 import PropTypes from 'prop-types';
 
 export default class ColumnGroup extends Component {
+  static isTableColumnGroup = true;
   static propTypes = {
     title: PropTypes.node,
   };
-
-  static isTableColumnGroup = true;
 }