+++ src/Table.js
@@ -6,7 +6,7 @@ import { Provider, create } from 'mini-store';
 import merge from 'lodash/merge';
 import classes from 'component-classes';
 import { polyfill } from 'react-lifecycles-compat';
-import { debounce, warningOnce } from './utils';
+import { debounce, warningOnce, getDataAndAriaProps } from './utils';
 import ColumnManager from './ColumnManager';
 import HeadTable from './HeadTable';
 import BodyTable from './BodyTable';
@@ -75,21 +75,6 @@ class Table extends React.Component {
     emptyText: () => 'No Data',
   };
 
-  static getDerivedStateFromProps(nextProps, prevState) {
-    if (nextProps.columns && nextProps.columns !== prevState.columns) {
-      return {
-        columns: nextProps.columns,
-        children: null,
-      };
-    } else if (nextProps.children !== prevState.children) {
-      return {
-        columns: null,
-        children: nextProps.children,
-      };
-    }
-    return null;
-  }
-
   constructor(props) {
     super(props);
 
@@ -113,7 +98,7 @@ class Table extends React.Component {
     this.store = create({
       currentHoverKey: null,
       fixedColumnsHeadRowsHeight: [],
-      fixedColumnsBodyRowsHeight: [],
+      fixedColumnsBodyRowsHeight: {},
     });
 
     this.setScrollPosition('left');
@@ -149,10 +134,33 @@ class Table extends React.Component {
     };
   }
 
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (nextProps.columns && nextProps.columns !== prevState.columns) {
+      return {
+        columns: nextProps.columns,
+        children: null,
+      };
+    } else if (nextProps.children !== prevState.children) {
+      return {
+        columns: null,
+        children: nextProps.children,
+      };
+    }
+    return null;
+  }
+
   componentDidMount() {
     if (this.columnManager.isAnyColumnsFixed()) {
       this.handleWindowResize();
       this.resizeEvent = addEventListener(window, 'resize', this.debouncedWindowResize);
+    }
+
+    // https://github.com/ant-design/ant-design/issues/11635
+    if (this.headTable) {
+      this.headTable.scrollLeft = 0;
+    }
+    if (this.bodyTable) {
+      this.bodyTable.scrollLeft = 0;
     }
   }
 
@@ -244,11 +252,18 @@ class Table extends React.Component {
       headRows,
       row => row.getBoundingClientRect().height || 'auto',
     );
-    const fixedColumnsBodyRowsHeight = [].map.call(
-      bodyRows,
-      row => row.getBoundingClientRect().height || 'auto',
-    );
     const state = this.store.getState();
+    const fixedColumnsBodyRowsHeight = [].reduce.call(
+      bodyRows,
+      (acc, row) => {
+        const rowKey = row.getAttribute('data-row-key');
+        const height =
+          row.getBoundingClientRect().height || state.fixedColumnsBodyRowsHeight[rowKey] || 'auto';
+        acc[rowKey] = height;
+        return acc;
+      },
+      {},
+    );
     if (
       shallowequal(state.fixedColumnsHeadRowsHeight, fixedColumnsHeadRowsHeight) &&
       shallowequal(state.fixedColumnsBodyRowsHeight, fixedColumnsBodyRowsHeight)
@@ -487,6 +502,7 @@ class Table extends React.Component {
     }
     const hasLeftFixed = this.columnManager.isAnyColumnsLeftFixed();
     const hasRightFixed = this.columnManager.isAnyColumnsRightFixed();
+    const dataAndAriaProps = getDataAndAriaProps(props);
 
     return (
       <Provider store={this.store}>
@@ -499,6 +515,7 @@ class Table extends React.Component {
                 className={className}
                 style={props.style}
                 id={props.id}
+                {...dataAndAriaProps}
               >
                 {this.renderTitle()}
                 <div className={`${prefixCls}-content`}>