+++ src/utils.js
@@ -1,6 +1,7 @@
 import warning from 'warning';
 
-let scrollbarSize;
+let scrollbarVerticalSize;
+let scrollbarHorizontalSize;
 
 // Measure scrollbar width for padding body during modal show/hide
 const scrollbarMeasure = {
@@ -8,31 +9,40 @@ const scrollbarMeasure = {
   top: '-9999px',
   width: '50px',
   height: '50px',
-  overflow: 'scroll',
 };
 
 export function measureScrollbar(direction = 'vertical') {
   if (typeof document === 'undefined' || typeof window === 'undefined') {
     return 0;
   }
-  if (scrollbarSize) {
-    return scrollbarSize;
+  const isVertical = direction === 'vertical';
+  if (isVertical && scrollbarVerticalSize) {
+    return scrollbarVerticalSize;
+  } else if (!isVertical && scrollbarHorizontalSize) {
+    return scrollbarHorizontalSize;
   }
   const scrollDiv = document.createElement('div');
   Object.keys(scrollbarMeasure).forEach(scrollProp => {
     scrollDiv.style[scrollProp] = scrollbarMeasure[scrollProp];
   });
+  // Append related overflow style
+  if (isVertical) {
+    scrollDiv.style.overflowY = 'scroll';
+  } else {
+    scrollDiv.style.overflowX = 'scroll';
+  }
   document.body.appendChild(scrollDiv);
   let size = 0;
-  if (direction === 'vertical') {
+  if (isVertical) {
     size = scrollDiv.offsetWidth - scrollDiv.clientWidth;
-  } else if (direction === 'horizontal') {
+    scrollbarVerticalSize = size;
+  } else {
     size = scrollDiv.offsetHeight - scrollDiv.clientHeight;
+    scrollbarHorizontalSize = size;
   }
 
   document.body.removeChild(scrollDiv);
-  scrollbarSize = size;
-  return scrollbarSize;
+  return size;
 }
 
 export function debounce(func, wait, immediate) {
@@ -78,4 +88,17 @@ export function remove(array, item) {
   const front = array.slice(0, index);
   const last = array.slice(index + 1, array.length);
   return front.concat(last);
+}
+
+/**
+ * Returns only data- and aria- key/value pairs
+ * @param {object} props
+ */
+export function getDataAndAriaProps(props) {
+  return Object.keys(props).reduce((memo, key) => {
+    if (key.substr(0, 5) === 'data-' || key.substr(0, 5) === 'aria-') {
+      memo[key] = props[key];
+    }
+    return memo;
+  }, {});
 }