+++ tests/Table.expandRow.spec.js
@@ -81,6 +81,27 @@ describe('Table.expand', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
+  /* eslint-disable react/prop-types */
+  it('renders a custom icon', () => {
+    function CustomExpandIcon(props) {
+      return (
+        <a className="expand-row-icon" onClick={e => props.onExpand(props.record, e)}>
+          <i className="some-class" />
+        </a>
+      );
+    }
+    const wrapper = render(
+      createTable({
+        expandedRowRender,
+        expandIcon: ({ onExpand, record }) => (
+          <CustomExpandIcon onExpand={onExpand} record={record} />
+        ),
+      }),
+    );
+    expect(wrapper).toMatchSnapshot();
+  });
+  /* eslint-enable react/prop-types */
+
   it('renders nested data correctly', () => {
     const localData = [
       {
@@ -204,5 +225,15 @@ describe('Table.expand', () => {
       .first()
       .simulate('click');
     expect(wrapper.render()).toMatchSnapshot();
+  });
+
+  // https://github.com/ant-design/ant-design/issues/12208
+  it('avoid onExpandedRowsChange repeat trigger when rows removed', () => {
+    const onExpandedRowsChange = jest.fn();
+    const wrapper = mount(createTable({ expandedRowRender, onExpandedRowsChange }));
+    expect(onExpandedRowsChange).not.toBeCalled();
+    wrapper.setProps({ data: [{ key: 93, name: 'Bamboo', age: 14 }] });
+    expect(onExpandedRowsChange).toBeCalledWith([]);
+    expect(onExpandedRowsChange.mock.calls.length).toEqual(1);
   });
 });